Revert "- let gl_draw_sync default to false."
This reverts commit bce29820493f450863cdbc5703f38bf45427307b.

Now I remember why this is not active: This is causing some problems with interpolated plane heights. Since there is no synchronization with the vertex buffer, the next frame will reset all plane coordinates before the GPU can use them. This will require some rethinking about when to update the buffer - it cannot be done in the render data setup loop.