- removed support for loading Build palettes as the main game palette.
This was implemented in a way that made it entirely impossible to load Build resources and make them usable for modding.
ZDoom had Build texture support for many years but the limitations the palette handling imposed made it impossible to use them.
It wasn't usable for anything more than to load Build maps and have them display properly - a feature that had to be removed because it was irreparably broken already.

With the forced palette override out of the way it should now be possible to implement loading of Build ART files as actually usable resources.